my_data_app/app/requirements.txt
pandas==1.3.3
google-cloud-bigquery==2.24.0